removed unnecessary code duplicate
